TreeBuilder: fix x_get_tree_ems_kids.
